requirements.txt
pyarrow
pandas
numpy
pillow
datasets
transformers
torch
torchvision
tqdmaccelerate==1.12.0
tensorboardX==2.6.4
